perf(core): pack the device registry record
A `DeviceListRecord` is held per known contact for the life of a cache entry,
and it spent most of itself on padding and duplication.

`DeviceInfo` carried a `u32` device id, an `Option<u32>` key index and a
`bool`: 5 bytes of information in 16, the rest alignment padding. Packed to 8 —
the device id is a `u16` because that is what it is on the wire and what
`Jid::device` has always been, and the hosted flag shares a byte with the key
index's presence bit. `Some(0)` and `None` stay distinct, which a sentinel
value would not have managed.

`user` was a `String` holding exactly what the registry cache keys the record
by: `update_device_list_guarded` assigns the canonical name to both, so every
cached user paid for two allocations of the same bytes. It is an `Arc<str>`
now, and the cache is keyed by the same one.

`devices` becomes a `Box<[DeviceInfo]>` and `phash` an `Option<Box<str>>`. The
list is built once and read for the life of the entry, so the capacity field is
dead weight — and `retain_devices_by_key_index` shortens it without releasing
capacity, leaving the filtered devices' slots resident. Mutations (all of them
notification paths) go through `edit_devices`.

Measured over 2000 contacts at 3 devices each: 261 -> 160 bytes per contact,
and 2 allocations instead of 4. `size_of::<DeviceListRecord>()` 88 -> 64.

Persisted blobs are unchanged in both directions: `DeviceInfo` serializes
through a shadow struct with the same three fields, `is_hosted` still defaults
when absent, and a blob claiming a device id past `u16` is now rejected by
serde instead of being truncated into a different device — which also removes
the two defensive `u16::try_from` paths that used to guard for it, since the
type no longer admits the value.

+++ src/retry.rs
@@ -522,7 +522,7 @@ impl Client {
         // list for this user, so refresh it (rate-limited, dedup'd) to learn the
         // device for the next send. Done before the message-cache lookup so an
         // evicted retry still triggers it.
-        let sender_device_id = info.requester.device() as u32;
+        let sender_device_id = info.requester.device();
         let device_known = self
             .has_device(&info.requester.user, sender_device_id)
             .await;
